01 - Form
We just started the course and you already devoured Lesson 1? Damn it! Before moving on, shall we review what we saw? You learned to:

Allow app interaction with form widget:
We create the base of the form screen with containerand column. We also “beautify” our canvas with color, BorderRadiusand border.

Using text fields with TextField:
Referring to the Cookbook , we create and style a text field where the app user can enter information. We have seen that there are two types: TextField and TextFormField.

Implement text fields with the TextFormField:
To allow the creation of new tasks in the app, we used and styled the TextFormField to implement text fields where the user can enter information;

We have seen that the differential of TextFormFieldis the parameter validator(Validation in Portuguese). For example, in the text field intended for difficulty, we want to ensure that the user enters numbers; validatorhelps us to validate this entered information .

Manage information with Controller:
In order to manage and store the information of the text fields, we use a parameter (do TextFormField) called Controller. To operate the Controller, we created a variable, called it no Controllerand tested it with print to verify that everything was correct with the storage of the information typed in the form's text fields.

We did the “conversion” from string to integer with the ìnt.parse` function, allowing the manipulation of numbers in the difficulty field (controlled by difficultyController.text).

Insert an image into the task:
We plan for the user to add an image that is shown in real time on the form! We did this with a Containerstyled one and its child ( child) Image.network(imageController.text), which basically pulls an image from our text field. To display the image in real time, we use the setStatewithin the TextFormField.
Handling errors with ErrorBuilder

We start taking care of the boogeymen! With ErroBuilder, which keeps an eye out for exceptions, we have created a solution in the case of invalid images! On error (if there is no image or it is invalid), ErrorBuilderconstructs (returns) a still image which indicates that there are no images. Thus, our app is elegant and informs something useful to the user. Too chic, right?