02 - Validation and Keyboard
We learned more app building skills with Flutter. Shall we review?

In this class, you learned to:
Avoid overflow error with SingleChildScrollView:
We prevent the phone keyboard from hiding parts of the screen. As? We wrap the 'Container (responsável por toda tela) com um novo Widget: oSingleChildScrollView`. Thus, it is possible to scroll and navigate the screen, avoiding keyboard overlap.

Ease the user experience with KeyBoardType:
When the user fills in a “name” text field, it is best to invoke a letter (alphabetical) keyboard. With this logic in mind, we customized the keyboards called by each of the text fields: the “name” field calls a letter keyboard; the “difficulty” field invokes the numerical keyboard; finally, the “image” field calls a top keyboard for URLs.

Prevent empty information from being submitted by the form:
We have learned that it is bad practice to submit nulls in a form. We don't want to display, for example, a task with no title or other information missing. To avoid this, we use the validator- parameter of TextFormField. Inside the validator, we create a condition to check if the field contains information or a null.

Monitor validations with the widget Form:
We saw that validations need someone to take care of them. That's where the new widget comes in: Form! It is responsible for verifying validations in a simpler way.

Show the user that the information was sent with the Snackbar:
To give more clarity about the app, we created a bar at the bottom of the screen that informs the user that the information entered in the form is correct and being saved. A snackbarbelongs to Scaffoldand appears only when all validations succeed (and disappears automatically)!