03 - Navigation and Routes
We're halfway there! In this class, you learned to:

Understand the concepts of navigation and routes:
- We learned two important concepts: navigation and routes . In the real world, navigation is the act of going from one screen to another; route is the path from one place to another. In Flutter, the widget Navigatorallows navigation between two screens. , routes(“routes”) are the screens that will be changed during navigation.
Identify the main types of navigation:

We know the fundamentals of navigation in Flutter:
- Navigation for apps: Involves a single path both outbound and back, one screen at a time. Also, this type of navigation is through stacks and takes place in a model called imperative, more linear and rigid;
- Navigation on websites: it allows for different paths, it is more dynamic, there is more freedom, it follows the declarative model - more complex. It's called Navigation 2.0;
- Anonymous routes: the simplest way to go from one screen to another in Flutter, use the widget Navigatorto add or remove a screen from the stack;
- Named Routes: Gives names to screens, organizing navigation and is best applied to an app that has many screens. Setup is more work.

- Apply navigation between two screens in an app:
We applied the navigation from the home screen to the form screen. First, we create a button ( FloatingActionButton) that activates Navigator.pushto create the route between screens and add a screen to the stack; then, on the form screen, inside the “add” button ( ElevatedButton), we use one navigator.pop to take the form screen off the stack, returning the app to the initial screen. Thus, when the user presses the “add” button on the form screen, it will return to the home page.