04 - Conversation between widgets
In this class, you learned to:

- Understand the concept of state in Flutter:
We saw that, in order to improve the app, it was necessary to understand the concept of state. Like a full or empty mug, variables and objects vary: they have different states. For example, a string can have three states - filled; empty; or null (three states). So, in the code, someone needs to keep an eye on information that is changing, which is done by setState, for example.

- Distinguish between ephemeral states and application states:*
We learned the difference between ephemeral state ( ephemeral states that concern a widget) and app state (in Portuguese, application states that require more care, as the data belongs to more than one widget).

- Pass information between widgets with InheritedWidget:
One more Widget for the account! is InheritedWidgetlike a delivery guy - it just receives data and passes it from one widget to another (it's like the “back-end” of the app). It is essential to pass information from the form screen to the initial screen, showing the created task. To allow for this, we've refactored the code responsible for our to-do list! In addition, we saw that it was necessary to configure and access InheritedWidget, making it the parent of all widgets.

- Understand the concept of context:
In short, context is where information is located. For example, the splash screen is a context; the form canvas is somewhere else (context). To guide yourself, you can use the tree of widgets to find out in which context information is. Finally, we delve into the of(Context).

- Save a task to the Home screen list:
After understanding the concept of context, we implemented the functionality of saving a new task on the app's home screen! Very interesting, right?