05 - Refining our project In this class, you learned to:
Fix task image bug:
Oh no! The app was almost finished, but the task image didn't appear on the home screen! Soon, you learned how to handle this bug. As? We investigated the code and found that our task only accepted images of the Asstes type, to resolve this we added a ternary operator with assetOrNetworkwhich it checks whether the image information is of the Network type or the Asset type, hence our project knows how to choose which Widget it will use to display the image.

Increase code safety with NullSafety
At the end of a project, it is good practice to review the code, test the project, and refactor points that can be improved. Some of these points were related to the TextFormField; in one of them, we ignore the possibility of the value being null - which can cause a big problem. So we fixed those validations and added a Null Safety for security.

Solve the bug that reset the task level
When going up or down the screen, the app resets the task level information! How to prevent this? We placed the int level above the @override, avoiding the reset of the state when the information “leaves” the screen. We've also made adjustments to all the code points where the level appears.